+++ docs/conf.py
@@ -34,7 +34,8 @@ class MMock(MagicMock):
 
 
 MOCK_MODULES = ['scipy', 'sklearn', 'clipt', 'skyfield', 'matplotlib', 'shapely',
-                'craytraverse', 'craytraverse.craytraverse', 'craytraverse.crenderer',
+                'shapely.ops', 'craytraverse', 'craytraverse.craytraverse',
+                'craytraverse.crenderer',
                 'craytraverse.crenderer.rcontrib_c', 'sklearn.cluster',
                 'craytraverse.crenderer.rtrace_c', 'scipy.ndimage',
                 'scipy.ndimage.filters', 'scipy.sparse', 'scipy.spatial',