+++ raytraverse/lightfield/sunsplanekd.py
@@ -8,7 +8,6 @@
 import numpy as np
 from scipy.spatial import cKDTree, distance_matrix
 
-from raytraverse import translate
 from raytraverse.lightfield.sets import MultiLightPointSet
 from raytraverse.lightfield.lightfield import LightField
 from raytraverse.translate import theta2chord, chord2theta
@@ -26,30 +25,68 @@ class SunsPlaneKD(LightField):
 
     @vecs.setter
     def vecs(self, pt):
+        if pt is None:
+            pt = f"{self.scene.outdir}/{self.pm.name}/{self.src}positions.tsv"
         pts, idx, samplelevel = self._load_vecs(pt)
-        # calculate sun sampling resolution for weighting query vecs
-        s0 = pts[samplelevel == 0]
-        dm = distance_matrix(s0, s0)
-        cm = np.ma.MaskedArray(dm, np.eye(dm.shape[0]))
-        sund = np.average(np.min(cm, axis=0).data)
-        self._normalization = sund / (self.pm.ptres / 2 * 2**.5)
-        s_pts = []
-        s_idx = []
-        s_lev = []
-        for i, pt, sl in zip(idx, pts, samplelevel):
-            source = f"{self.src}_{i:04d}"
-            ptf = f"{self.scene.outdir}/{self.pm.name}/{source}_points.tsv"
-            spt, sidx, slev = self._load_vecs(ptf)
-            s_pts.append(np.hstack((np.broadcast_to(pt[None], spt.shape), spt)))
-            s_idx.append(np.stack((np.broadcast_to([i], sidx.shape), sidx)).T)
-            s_lev.append(np.stack((np.broadcast_to([sl], slev.shape), slev)).T)
-        self._vecs = np.concatenate(s_pts)
+        s_pts, s_lev, s_idx, nmz = self._load_points(idx, pts, samplelevel)
+        self._normalization = nmz
+        self._vecs = s_pts
+        self._samplelevel = s_lev
+        self.data = s_idx
         self._kd = None
         self._suns = pts
         self._sunkd = None
-        self._samplelevel = np.concatenate(s_lev)
         self.omega = None
-        self.data = np.concatenate(s_idx)
+
+    def _load_points(self, idx, pts, samplelevel):
+        """loop over source point files to build indices"""
+        s_pts = []
+        s_idx = []
+        s_lev = []
+        pdist = []
+        # fall back to guesstimate of level 0 in case component planes
+        # do not have levels indexed
+        l0 = self.pm.point_grid(False).shape[0]
+        haslevels = np.sum(samplelevel) > 0
+        if haslevels:
+            # calculate sun sampling resolution for weighting query vecs
+            sund = self._estimate_grid_size(pts[samplelevel == 0])
+        else:
+            sund = self._estimate_grid_size(pts[:45])
+
+        for i, pt, sl in zip(idx, pts, samplelevel):
+            spt, sidx, slev = self._load_point(f"{self.src}_{i:04d}")
+            # only the unmasked level 0 suns will provide a good
+            # estimate for level0 point distances
+            if haslevels and sl == 0:
+                if np.sum(slev) > 0:
+                    s = spt[slev == 0]
+                else:
+                    s = spt[:l0]
+                pdist.append(self._estimate_grid_size(s))
+
+            s_pts.append(np.hstack((np.broadcast_to(pt[None], spt.shape), spt)))
+            s_idx.append(np.stack((np.broadcast_to([i], sidx.shape), sidx)).T)
+            s_lev.append(np.stack((np.broadcast_to([sl], slev.shape), slev)).T)
+        if haslevels:
+            normalization = sund/np.average(pdist)
+        else:
+            pdist = self.pm.ptres * (np.sqrt(2)/2+.5)
+            normalization = sund/pdist
+        s_pts = np.concatenate(s_pts)
+        s_lev = np.concatenate(s_lev)
+        s_idx = np.concatenate(s_idx)
+        return s_pts, s_lev, s_idx, normalization
+
+    def _load_point(self, source):
+        ptf = f"{self.scene.outdir}/{self.pm.name}/{source}_points.tsv"
+        return self._load_vecs(ptf)
+
+    @staticmethod
+    def _estimate_grid_size(v):
+        dm = distance_matrix(v, v)
+        cm = np.ma.MaskedArray(dm, np.eye(dm.shape[0]))
+        return np.average(np.min(cm, axis=0).data)
 
     @property
     def suns(self):
@@ -95,15 +132,15 @@ class SunsPlaneKD(LightField):
             integer indices of closest ray to each query
         d: np.array
             distance from query to point, positional distance is normalized
-            by the average chord-length between leveel 0 sun samples divided by
-            the PlanMapper ptres * sqrt(2).
+            by the average chord-length between level 0 sun samples divided by
+            the average distance between level 0 pt samples.
         """
         weighted_vecs = np.copy(np.atleast_2d(vecs)).astype(float)
         weighted_vecs[:, 3:] *= self._normalization
         d, i = self.kd.query(weighted_vecs)
         return i, d
 
-    def query_by_sun(self, sunvec, fixed_points=None, ptfilter=.25, stol=10,
+    def query_by_sun(self, sunvec, fixed_points=None, stol=10,
                      minsun=1):
         """for finding vectors across zone, sun vector based query
 
@@ -115,8 +152,6 @@ class SunsPlaneKD(LightField):
             2d array like, shape (N, 3) of additional fixed points to return
             use for example with a matching sky query. Note that if point filter
             is to large not all of these points are necessarily returned.
-        ptfilter: Union[float, int], optional
-            minimum seperation for returned points
         stol: Union[float, int], optional
             maximum angle (in degrees) for matching sun vectors
         minsun: int, optional
@@ -136,10 +171,11 @@ class SunsPlaneKD(LightField):
 
         """
         return _query_by_sun(sunvec, self.sunkd, self.vecs, self.data.idx[:, 0],
-                             fixed_points=fixed_points, ptfilter=ptfilter,
-                             stol=stol, minsun=minsun)
+                             fixed_points=fixed_points,
+                             stol=stol, minsun=minsun,
+                             normalization=self._normalization)
 
-    def query_by_suns(self, sunvecs, fixed_points=None, ptfilter=.25, stol=10,
+    def query_by_suns(self, sunvecs, fixed_points=None, stol=10,
                       minsun=1):
         """parallel processing call to query_by_sun for 2d array of sunvecs
 
@@ -151,8 +187,6 @@ class SunsPlaneKD(LightField):
             2d array like, shape (N, 3) of additional fixed points to return
             use for example with a matching sky query. Note that if point filter
             is to large not all of these points are necessarily returned.
-        ptfilter: Union[float, int], optional
-            minimum seperation for returned points
         stol: Union[float, int], optional
             maximum angle (in degrees) for matching sun vectors
         minsun: int, optional
@@ -173,16 +207,20 @@ class SunsPlaneKD(LightField):
                            self.data.idx[:, 0], expandarg=False,
                            desc="finding SunPlane points",
                            pbar=self.scene.dolog, fixed_points=fixed_points,
-                           ptfilter=ptfilter, stol=stol, minsun=minsun)
+                           stol=stol, minsun=minsun,
+                           normalization=self._normalization)
         vecs, idx, d = zip(*result)
         return vecs, idx, d
 
 
 def _query_by_sun(sunvec, sunkd, svecs, sidx, fixed_points=None,
-                  ptfilter=.25, stol=10, minsun=1):
+                  stol=10, minsun=1, normalization=1.0):
     """for finding vectors across zone, sun vector based query"""
-    # find suns within tolerance or minimum number
-    i = sunkd.query_ball_point(np.ravel(sunvec), theta2chord(stol*np.pi/180))
+
+    # first find all points with suns within tolerance
+    cstol = theta2chord(stol*np.pi/180)
+    i = sunkd.query_ball_point(np.ravel(sunvec), cstol)
+
     if len(i) < minsun:
         d, i = sunkd.query(sunvec, minsun)
         if minsun == 1:
@@ -192,8 +230,7 @@ def _query_by_sun(sunvec, sunkd, svecs, sidx, fixed_points=None,
     sunfilt = np.isin(sidx, i)
     vecs = svecs[sunfilt]
     idx = np.argwhere(sunfilt).ravel()
-
-    # prepend fixed_points
+    # prepend fixed_points (using closest values)
     if fixed_points is not None:
         fixed_points = np.atleast_2d(fixed_points)
         pkd = cKDTree(vecs[:, 3:])
@@ -202,13 +239,14 @@ def _query_by_sun(sunvec, sunkd, svecs, sidx, fixed_points=None,
         vecs = np.concatenate((fvecs, vecs), axis=0)
         idx = np.concatenate((idx[i], idx))
 
-    # sort points by ascending by sunerror
+    # calculate sun errors
     sund = np.linalg.norm(vecs[:, 0:3] - np.atleast_2d(sunvec), axis=1)
-    sorti = np.argsort(sund, kind='stable')
+    cpts = np.copy(vecs[:, 3:])
+    # add sun error penalty
+    cpts[:, 2] += sund/normalization
+    ckd = cKDTree(cpts)
+    # reassign original vectors based on penalized query
+    d, i = ckd.query(vecs[:, 3:])
 
-    # cull points within tolerance prioritizing minimum sun error
-    flt = translate.cull_vectors(vecs[sorti, 3:], ptfilter)
-    # sort back to original order and return indices with sun errors
-    flt = flt[np.argsort(sorti, kind='stable')]
-    d = chord2theta(sund[flt]) * 180/np.pi
-    return vecs[flt], idx[flt], d
+    d = chord2theta(sund[i]) * 180/np.pi
+    return vecs[i], idx[i], d